Controller pin layout
This file contains all of the joystick buttons mapped to the cabinet as a whole. I have shared it as it may help provide context to the buttons mapped in the Arduino code. The excel doc also includes an image of the PCB designed for the micro. Its a simple design with a few pull down resistors and some protection for the LED strip DIO pins.